+++ services/web/app/storage/routes/site.py
@@ -11,8 +11,6 @@ from ..models import (
     Room,
     FixedColdStorage,
     FixedColdStorageShelf,
-    SampleToFixedColdStorageShelf,
-    CryovialBoxToFixedColdStorageShelf,
 )
 
 from ...misc.models import Address
@@ -117,7 +115,7 @@ def get_data(id):
                     "samples": {},
                     "cryo": {},
                 }
-
+                '''
                 samples_to_shelf = (
                     db.session.query(SampleToFixedColdStorageShelf)
                     .filter(SampleToFixedColdStorageShelf.shelf_id == shelf.id)
@@ -140,6 +138,7 @@ def get_data(id):
                     output[room.id]["storage"][storage.id]["shelves"][shelf.id]["cryo"][
                         cryo.id
                     ] = {"test": "data"}
+                '''
 
     return jsonify(output), 201, {"Content-Type": "application/json"}
