+++ services/web/app/storage/views/cryobox.py
@@ -1,5 +1,5 @@
 from ... import db
-from ..models import CryovialBox, SampleToCryovialBox
+from ..models import CryovialBox
 from ...auth.views import UserView
 from ...sample.views import BasicSampleView
 
@@ -32,8 +32,8 @@ def CryoboxView(cryo_id: int) -> dict:
 
     data["info"] = BasicCryoboxView(cryo_id)
     data["sample_information"] = {}
-
+    '''
     for sample in db.session.query(SampleToCryovialBox).filter(SampleToCryovialBox.box_id == cryo_id).all():
         data["sample_information"]["%i_%i" % (sample.row, sample.col)] = BasicSampleView(sample.sample_id)
-
+    '''
     return data
